fix golint && uint32

+++ carbonserver/carbonserver.go
@@ -130,7 +130,7 @@ type ExpandedGlobResponse struct {
 	Files     []string
 	Leafs     []bool
 	TrieNodes []*trieNode
-	Lookups   uint64
+	Lookups   uint32
 	Err       error
 }
